rm integration_utils crate
this crate was keeping `integration_check` an std crate, causing it to fail to build on windows and some ubuntu setups
