--- core/rs/integration-utils/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "integration_utils"
-version = "0.0.1"
-edition = "2021"
-authors = ["Matt Wonlaw"]
-description = "code shared between integration tests"
-keywords = ["sqlite"]
-license = "Apache-2.0"
-
-[dependencies]
-sqlite_nostd = { path="../sqlite-rs-embedded/sqlite_nostd", features=["static"] }
-
-[lib]
-name = "integration_utils"
-crate-type = ["rlib"]
