--- core/rs/integration-utils/rust-toolchain.toml
@@ -1,2 +1,0 @@
-[toolchain]
-channel = "nightly-2023-06-17"
